Separate main functionality, move drive callback into MBotApp object
